requirements.txt
channels==4.2.0
daphne==4.1.2
Django==5.1.7
djangorestframework==3.15.2
googlemaps==4.10.0
gunicorn==20.1.0
pillow==11.1.0
psycopg2==2.9.3
whitenoise==6.9.0